+++ test/generator/print_helpers_test.dart
@@ -474,6 +474,7 @@ import 'package:equatable/equatable.dart';
 import 'package:gql/ast.dart';
 part 'test_query.graphql.g.dart';
 
+final TEST_QUERY_QUERY_DOCUMENT_OPERATION_NAME = 'test_query';
 final TEST_QUERY_QUERY_DOCUMENT = DocumentNode(definitions: [
   OperationDefinitionNode(
       type: OperationType.query,
@@ -490,13 +491,53 @@ class TestQueryQuery extends GraphQLQuery<TestQuery, JsonSerializable> {
   final DocumentNode document = TEST_QUERY_QUERY_DOCUMENT;
 
   @override
-  final String operationName = 'test_query';
+  final String operationName = TEST_QUERY_QUERY_DOCUMENT_OPERATION_NAME;
 
   @override
   List<Object?> get props => [document, operationName];
   @override
   TestQuery parse(Map<String, dynamic> json) => TestQuery.fromJson(json);
 }
+''');
+    });
+
+    test(
+        'When generateHelpers is false and generateQueries is true, an execute fn is generated.',
+        () {
+      final buffer = StringBuffer();
+      final definition = LibraryDefinition(
+        basename: r'test_query.graphql',
+        queries: [
+          QueryDefinition(
+            name: QueryName(name: 'test_query'),
+            operationName: 'test_query',
+            document: parseString('query test_query {}'),
+            generateHelpers: false,
+            generateQueries: true,
+          )
+        ],
+      );
+      final ignoreForFile = <String>[];
+
+      writeLibraryDefinitionToBuffer(buffer, ignoreForFile, definition);
+
+      expect(buffer.toString(), '''// GENERATED CODE - DO NOT MODIFY BY HAND
+// @dart = 2.12
+
+import 'package:json_annotation/json_annotation.dart';
+import 'package:equatable/equatable.dart';
+import 'package:gql/ast.dart';
+part 'test_query.graphql.g.dart';
+
+final TEST_QUERY_QUERY_DOCUMENT_OPERATION_NAME = 'test_query';
+final TEST_QUERY_QUERY_DOCUMENT = DocumentNode(definitions: [
+  OperationDefinitionNode(
+      type: OperationType.query,
+      name: NameNode(value: 'test_query'),
+      variableDefinitions: [],
+      directives: [],
+      selectionSet: SelectionSetNode(selections: []))
+]);
 ''');
     });
 
@@ -520,7 +561,7 @@ class TestQueryQuery extends GraphQLQuery<TestQuery, JsonSerializable> {
 
       writeLibraryDefinitionToBuffer(buffer, ignoreForFile, definition);
 
-      expect(buffer.toString(), '''// GENERATED CODE - DO NOT MODIFY BY HAND
+      expect(buffer.toString(), r'''// GENERATED CODE - DO NOT MODIFY BY HAND
 // @dart = 2.12
 
 import 'package:artemis/artemis.dart';
@@ -535,16 +576,17 @@ class TestQueryArguments extends JsonSerializable with EquatableMixin {
 
   @override
   factory TestQueryArguments.fromJson(Map<String, dynamic> json) =>
-      _\$TestQueryArgumentsFromJson(json);
+      _$TestQueryArgumentsFromJson(json);
 
   final Type? name;
 
   @override
   List<Object?> get props => [name];
   @override
-  Map<String, dynamic> toJson() => _\$TestQueryArgumentsToJson(this);
+  Map<String, dynamic> toJson() => _$TestQueryArgumentsToJson(this);
 }
 
+final TEST_QUERY_QUERY_DOCUMENT_OPERATION_NAME = 'test_query';
 final TEST_QUERY_QUERY_DOCUMENT = DocumentNode(definitions: [
   OperationDefinitionNode(
       type: OperationType.query,
@@ -561,7 +603,7 @@ class TestQueryQuery extends GraphQLQuery<TestQuery, TestQueryArguments> {
   final DocumentNode document = TEST_QUERY_QUERY_DOCUMENT;
 
   @override
-  final String operationName = 'test_query';
+  final String operationName = TEST_QUERY_QUERY_DOCUMENT_OPERATION_NAME;
 
   @override
   final TestQueryArguments variables;
@@ -619,11 +661,12 @@ class TestQueryArguments extends JsonSerializable with EquatableMixin {
         suffix: 'Query',
       );
 
-      final str =
-          generateQueryClassSpec(definition).map((e) => specToString(e)).join();
+      final str = specToString(generateQuerySpec(definition)) +
+          specToString(generateQueryClassSpec(definition));
 
-      expect(
-          str, '''final TEST_QUERY_QUERY_DOCUMENT = DocumentNode(definitions: [
+      expect(str,
+          r'''final TEST_QUERY_QUERY_DOCUMENT_OPERATION_NAME = 'test_query';
+final TEST_QUERY_QUERY_DOCUMENT = DocumentNode(definitions: [
   OperationDefinitionNode(
       type: OperationType.query,
       name: NameNode(value: 'test_query'),
@@ -638,7 +681,7 @@ class TestQueryQuery extends GraphQLQuery<TestQuery, TestQueryArguments> {
   final DocumentNode document = TEST_QUERY_QUERY_DOCUMENT;
 
   @override
-  final String operationName = 'test_query';
+  final String operationName = TEST_QUERY_QUERY_DOCUMENT_OPERATION_NAME;
 
   @override
   final TestQueryArguments variables;